feat: Phase 3 — i18n, school UX, photometry, ADES, calibration, toast notifications
- Full bilingual interface: 1007 PT-BR strings (100% finished), EN source fallback
- School workflow: step-by-step guided panel, Export/Import School Config, Send to Teacher
- Toast notification widget (non-blocking success/info/warning/error)
- FAQ and For Students sections in built-in help (EN + PT-BR)
- Credits & License section in help with GitHub links to reference projects
- PolyForm Noncommercial 1.0.0 license (replaces MIT); copyright Petrus Silva Costa
- README, CONTRIBUTING, CHANGELOG (bilingual in single files)
- AboutDialog: PolyForm license dialog, clickable GitHub links for all reference projects
- Centroid and aperture photometry engine (Photometry, Centroid)
- Moving object detector with blink-based tracklet linking
- Image stacker (mean/median/sigma-clipped)
- Calibration wizard and panel (bias/dark/flat)
- ASTAP offline plate solver support
- MPC/ADES report generation and preview dialog
- Observatory preset database (MPC codes), MPCORB offline catalog (KOO engine)
- Session comparator, light curve dialog, growth curve dialog
- FITS project save/load (GusProject), image export
- Dynamic WCS status in workflow panel
- Day/Night/Auto theme toggle; complete settings dialog
- .gitignore: exclude lint artifacts, .claude/, semelhantes/

+++ src/core/StarDetector.cpp
@@ -81,7 +81,7 @@ detectStars(const FitsImage& img, const StarDetectorConfig& cfg)
     exIm.gain        = img.gain;
 
     // ── 5. Extract sources ────────────────────────────────────────────────
-    const float  thresh    = cfg.threshold;    // SEP_THRESH_REL = in sigma
+    const float  thresh    = static_cast<float>(cfg.threshold);    // SEP_THRESH_REL = in sigma
     const float* conv      = cfg.filterKernel ? kConv3x3 : nullptr;
     const int    convw     = cfg.filterKernel ? 3 : 0;
     const int    convh     = cfg.filterKernel ? 3 : 0;
